fix(deepseek-v4): attn_sink FP32 contract — kernels index const float*, loader normalizes to FP32
The checkpoint ships layers.N.attn.attn_sink as F32 [num_q_heads], but the DS4F
attention kernels declared/indexed it as const __nv_bfloat16* (2-byte stride over
a 4-byte buffer). Even head 2k read the low-mantissa half of fp32[k]; 7 reads were
large-positive and collapsed the softmax value accumulator to exact zero
(heads {6,10,12,20,28,34,48}); the surviving even heads were silently perturbed.

Canonical contract: MlaWeights.attn_sink is ALWAYS a device FP32 [num_q_heads]
buffer; every sink kernel indexes const float*.

Kernels (const __nv_bfloat16* sinks -> const float*; drop __bfloat162float / widening cast):
  inferspark_prefill_512, prefill_attn_compressed, mla_paged_decode_fp8,
  mla_paged_decode (mla_paged_decode_nvfp4).
Loader (assemble.rs): respect WeightTensor.dtype — F32 pass-through (byte no-op),
BF16 widened once to an owned FP32 buffer (process-lifetime, cf. load_hc_f32),
any other dtype fails loudly. Never index fp32 as bf16.
Tests: bf16 widening exact; fp32 read returns true sink; bf16 misread reproduces
the exact seven-head fingerprint.

Consumer wrappers pass a raw DevicePtr (u64) — ABI/arg order unchanged.
LOCAL-ONLY, no push/PR.

+++ kernels/gb10/deepseek-v4-flash/nvfp4/mla_paged_decode.cu
@@ -91,7 +91,7 @@ extern "C" __global__ void mla_paged_decode_nvfp4(
     const float inv_sqrt_d,                          // 1/sqrt(576)
     const unsigned long long block_stride_bytes,
     const unsigned long long data_section_bytes,
-    const __nv_bfloat16* __restrict__ sinks          // [num_q_heads] per-head attn sink (s_aux); may be NULL
+    const float* __restrict__ sinks                  // [num_q_heads] per-head attn sink (s_aux); may be NULL. FP32 checkpoint-native, indexed as float.
 ) {
     const unsigned int q_head = blockIdx.x;
     const unsigned int seq_idx = blockIdx.y;
@@ -346,7 +346,7 @@ extern "C" __global__ void mla_paged_decode_nvfp4(
         // Mirrors the FP8 MLA decode kernel (mla_paged_decode_fp8.cu) and eager_attention_forward.
         // smem_m[0] is the global running max after the cross-warp reduction above.
         if (sinks != nullptr) {
-            final_l += __expf((float)sinks[q_head] - smem_m[0]);
+            final_l += __expf(sinks[q_head] - smem_m[0]);
         }
         float inv_l = (final_l > 0.0f) ? (1.0f / final_l) : 0.0f;
         unsigned int* o32 = (unsigned int*)(O + (unsigned long long)q_head * q_head_dim + vec_offset_bf16);
