+++ kernels/gb10/deepseek-v4-flash/nvfp4/mla_paged_decode_fp8.cu
@@ -47,7 +47,7 @@ extern "C" __global__ void mla_paged_decode_fp8(
     const float k_scale,                             // FP8 scale for K
     const float v_scale,                             // FP8 scale for V
     const unsigned long long cache_stride_bytes,
-    const __nv_bfloat16* __restrict__ sinks          // [num_q_heads] per-head attn sink (s_aux); may be NULL
+    const float* __restrict__ sinks                  // [num_q_heads] per-head attn sink (s_aux); may be NULL. FP32 checkpoint-native, indexed as float.
 ) {
     const unsigned int q_head = blockIdx.x;
     const unsigned int seq_idx = blockIdx.y;
@@ -319,7 +319,7 @@ extern "C" __global__ void mla_paged_decode_fp8(
         // eager_attention_forward concats `sinks`, softmaxes, then slices off the
         // sink column). Online-softmax: add exp(sink - running_max) to the sum.
         if (sinks != nullptr) {
-            final_l += __expf((float)sinks[q_head] - smem_m[0]);
+            final_l += __expf(sinks[q_head] - smem_m[0]);
         }
         float inv_l = (final_l > 0.0f) ? (1.0f / final_l) : 0.0f;
         unsigned int* o32 = (unsigned int*)(O + (unsigned long long)q_head * q_head_dim + vec_offset_bf16);
